Do not store socket_addr and slave_id in Client
Only connect_tcp uses these variables. After connecting all that's
needed for communication with the inverter is the Context.

By not storing the socket_addr and slave_id, the Client struct is also
usable for Modbus-RTU connections (even though there is no helper
to set up the context -- the caller can do this on their own).