Add the ability to query multiple devices using the same Client
This change makes it so you can query multiple SunSpec devices on a single
connection, removing the overhead of reconnecting (and re-scanning for
available models) and/or keeping multiple connections active at the same
time.

Multiple inverters can be connected together using RS-485, or on the same
TCP connection. For instance, APSystems micro-inverters each show up on
separate addresses on the same Modbus instance on the ECU box.